初步完成报告并融合stride
初步完成报告，并实现stride比较，但并未shijiyingy

尝试融合stride，但是有bug

修复bug
成功将不怕溢出的stride加入进代码

+++ os/src/config.rs
@@ -24,5 +24,5 @@ pub const CLOCK_FREQ: usize = 12500000;
 /// the physical memory end
 pub const MEMORY_END: usize = 0x88000000;
 ///lab5 StrideInfo
-pub const INIT_PRIORITY:isize = 16;
-pub const BIGSTRDE:isize = isize::MAX >> 20; 
+pub const INIT_PRIORITY:u64 = 16;
+pub const BIGSTRDE:u64 = u64::MAX >> 20;
